Overview of matching quality

The overview function provides a quick assessment of overall matching quality by reporting the N, mean, and s.d. of focal, matched, pool, and unmatched sets for all covariates as well as the propensity scores ('ps'). The mean and s.d. difference in focal - matched is also reported.

overview(x, digits = 2, ...)

# S4 method for Matched,numeric_OR_missing
overview(x, digits)

Arguments

x
Matched object

digits
Integer indicating the number of significant digits to be used. Negative values are allowed (see ?signif).

...
Additional arguments.

Value

A printed overview of matching quality.

Details

Factor, character, or logical covariates are reported by N per set, rather than with mean and s.d.

Examples

mdf <- makeExampleMatchedDataSet(matched = TRUE)
overview(mdf)

#> MatchedDataFrame object: 
#>        set     N feature2.mean feature2.sd feature3.a feature3.b feature3.c
#>      focal   500           4.1         1.9         66        157        206
#>    matched   500           4.5         2.8         56        141        239
#>       pool 10000           6.0         3.4       4248       3121       1117
#>  unmatched  9500           6.1         3.4       4192       2980        878
#>  feature3.d feature3.e ps.mean ps.sd
#>          49         22   0.100 0.076
#>          46         18   0.110 0.076
#>         992        522   0.045 0.051
#>         946        504   0.041 0.047
#> --------
#> focal - matched: 
#>  feature2.mean feature2.sd feature3.a feature3.b feature3.c feature3.d
#>          -0.44       -0.94         10         16        -33          3
#>  feature3.e ps.mean   ps.sd
#>           4 -0.0036 0.00035