Function for generating an example matchRanges or Matched dataset

This function will generate an example dataset as either 1) input for matchRanges() (when matched = FALSE) or 2) a Matched Object (when matched = TRUE).

makeExampleMatchedDataSet(
  type = "DataFrame",
  matched = FALSE,
  method = "rejection",
  replace = FALSE,
  ...
)

# S4 method for character_OR_missing,logical_OR_missing,character_OR_missing,logical_OR_missing
makeExampleMatchedDataSet(type, matched, method, replace)

Arguments

type
Character designating which type of dataset to make. options are one of 'data.frame', 'data.table', 'DataFrame', 'GRanges', or 'GInteractions'.

matched
TRUE/FALSE designating whether to generate a Matched dataset (matched = TRUE) or an input dataset for matchRanges() (matched = FALSE).

method
Character describing which matching method to use. Supported options are either 'nearest', 'rejection', or 'stratified'.

replace
TRUE/FALSE describing whether to select matches with or without replacement.

...
Additional arguments

Value

Returns an example Matched dataset or an example dataset for input to matchRanges().

Details

When matched = FALSE, the data returned contains 3 different features that can be subset to perform matching.

Examples

## Make examples for matchRanges() (i.e matched = FALSE)
makeExampleMatchedDataSet()

#> DataFrame with 10500 rows and 3 columns
#>        feature1  feature2    feature3
#>       <logical> <numeric> <character>
#> 1          TRUE   2.87905           c
#> 2          TRUE   3.53965           c
#> 3          TRUE   7.11742           c
#> 4          TRUE   4.14102           a
#> 5          TRUE   4.25858           c
#> ...         ...       ...         ...
#> 10496     FALSE   1.23578           b
#> 10497     FALSE   1.69671           a
#> 10498     FALSE   6.11140           a
#> 10499     FALSE   2.21657           d
#> 10500     FALSE   5.33003           b

head(makeExampleMatchedDataSet(type = 'data.frame', matched = FALSE))

#>   feature1 feature2 feature3
#> 1     TRUE 2.879049        c
#> 2     TRUE 3.539645        c
#> 3     TRUE 7.117417        c
#> 4     TRUE 4.141017        a
#> 5     TRUE 4.258575        c
#> 6     TRUE 7.430130        c

makeExampleMatchedDataSet(type = 'data.table', matched = FALSE)

#>        feature1 feature2 feature3
#>     1:     TRUE 2.879049        c
#>     2:     TRUE 3.539645        c
#>     3:     TRUE 7.117417        c
#>     4:     TRUE 4.141017        a
#>     5:     TRUE 4.258575        c
#>    ---                           
#> 10496:    FALSE 1.235781        b
#> 10497:    FALSE 1.696712        a
#> 10498:    FALSE 6.111404        a
#> 10499:    FALSE 2.216575        d
#> 10500:    FALSE 5.330029        b

makeExampleMatchedDataSet(type = 'DataFrame', matched = FALSE)

#> DataFrame with 10500 rows and 3 columns
#>        feature1  feature2    feature3
#>       <logical> <numeric> <character>
#> 1          TRUE   2.87905           c
#> 2          TRUE   3.53965           c
#> 3          TRUE   7.11742           c
#> 4          TRUE   4.14102           a
#> 5          TRUE   4.25858           c
#> ...         ...       ...         ...
#> 10496     FALSE   1.23578           b
#> 10497     FALSE   1.69671           a
#> 10498     FALSE   6.11140           a
#> 10499     FALSE   2.21657           d
#> 10500     FALSE   5.33003           b

makeExampleMatchedDataSet(type = 'GRanges', matched = FALSE)

#> GRanges object with 10500 ranges and 3 metadata columns:
#>           seqnames      ranges strand |  feature1  feature2    feature3
#>              <Rle>   <IRanges>  <Rle> | <logical> <numeric> <character>
#>       [1]     chr1       1-100      * |      TRUE   2.87905           c
#>       [2]     chr1       2-101      * |      TRUE   3.53965           c
#>       [3]     chr1       3-102      * |      TRUE   7.11742           c
#>       [4]     chr1       4-103      * |      TRUE   4.14102           a
#>       [5]     chr1       5-104      * |      TRUE   4.25858           c
#>       ...      ...         ...    ... .       ...       ...         ...
#>   [10496]     chr1 10496-10595      * |     FALSE   1.23578           b
#>   [10497]     chr1 10497-10596      * |     FALSE   1.69671           a
#>   [10498]     chr1 10498-10597      * |     FALSE   6.11140           a
#>   [10499]     chr1 10499-10598      * |     FALSE   2.21657           d
#>   [10500]     chr1 10500-10599      * |     FALSE   5.33003           b
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

makeExampleMatchedDataSet(type = 'GInteractions', matched = FALSE)

#> GInteractions object with 10500 interactions and 3 metadata columns:
#>           seqnames1     ranges1     seqnames2     ranges2 |  feature1  feature2
#>               <Rle>   <IRanges>         <Rle>   <IRanges> | <logical> <numeric>
#>       [1]      chr1       1-100 ---      chr1       1-100 |      TRUE   2.87905
#>       [2]      chr1       2-101 ---      chr1       2-101 |      TRUE   3.53965
#>       [3]      chr1       3-102 ---      chr1       3-102 |      TRUE   7.11742
#>       [4]      chr1       4-103 ---      chr1       4-103 |      TRUE   4.14102
#>       [5]      chr1       5-104 ---      chr1       5-104 |      TRUE   4.25858
#>       ...       ...         ... ...       ...         ... .       ...       ...
#>   [10496]      chr1 10496-10595 ---      chr1 10496-10595 |     FALSE   1.23578
#>   [10497]      chr1 10497-10596 ---      chr1 10497-10596 |     FALSE   1.69671
#>   [10498]      chr1 10498-10597 ---      chr1 10498-10597 |     FALSE   6.11140
#>   [10499]      chr1 10499-10598 ---      chr1 10499-10598 |     FALSE   2.21657
#>   [10500]      chr1 10500-10599 ---      chr1 10500-10599 |     FALSE   5.33003
#>              feature3
#>           <character>
#>       [1]           c
#>       [2]           c
#>       [3]           c
#>       [4]           a
#>       [5]           c
#>       ...         ...
#>   [10496]           b
#>   [10497]           a
#>   [10498]           a
#>   [10499]           d
#>   [10500]           b
#>   -------
#>   regions: 10500 ranges and 0 metadata columns
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Make Matched class examples (i.e. matched = TRUE)
makeExampleMatchedDataSet(matched = TRUE)

#> MatchedDataFrame with 500 rows and 3 columns
#>      feature1  feature2    feature3
#>     <logical> <numeric> <character>
#> 1       FALSE   5.60013           c
#> 2       FALSE   5.43395           c
#> 3       FALSE   4.47920           b
#> 4       FALSE   1.45446           c
#> 5       FALSE   2.99545           c
#> ...       ...       ...         ...
#> 496     FALSE   4.90340           c
#> 497     FALSE   9.54752           b
#> 498     FALSE   4.64822           c
#> 499     FALSE   1.82828           d
#> 500     FALSE   7.02681           b

makeExampleMatchedDataSet(type = 'DataFrame', matched = TRUE,
                          method = 'rejection',
                          replace = FALSE)

#> MatchedDataFrame with 500 rows and 3 columns
#>      feature1  feature2    feature3
#>     <logical> <numeric> <character>
#> 1       FALSE   5.60013           c
#> 2       FALSE   5.43395           c
#> 3       FALSE   4.47920           b
#> 4       FALSE   1.45446           c
#> 5       FALSE   2.99545           c
#> ...       ...       ...         ...
#> 496     FALSE   4.90340           c
#> 497     FALSE   9.54752           b
#> 498     FALSE   4.64822           c
#> 499     FALSE   1.82828           d
#> 500     FALSE   7.02681           b

makeExampleMatchedDataSet(type = 'GRanges', matched = TRUE,
                          method = 'rejection',
                          replace = FALSE)

#> MatchedGRanges object with 500 ranges and 3 metadata columns:
#>         seqnames      ranges strand |  feature1  feature2    feature3
#>            <Rle>   <IRanges>  <Rle> | <logical> <numeric> <character>
#>     [1]     chr1   4191-4290      * |     FALSE   5.60013           c
#>     [2]     chr1   3668-3767      * |     FALSE   5.43395           c
#>     [3]     chr1 10236-10335      * |     FALSE   4.47920           b
#>     [4]     chr1   8451-8550      * |     FALSE   1.45446           c
#>     [5]     chr1   5658-5757      * |     FALSE   2.99545           c
#>     ...      ...         ...    ... .       ...       ...         ...
#>   [496]     chr1   7472-7571      * |     FALSE   4.90340           c
#>   [497]     chr1   8316-8415      * |     FALSE   9.54752           b
#>   [498]     chr1   8987-9086      * |     FALSE   4.64822           c
#>   [499]     chr1   8288-8387      * |     FALSE   1.82828           d
#>   [500]     chr1     868-967      * |     FALSE   7.02681           b
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

makeExampleMatchedDataSet(type = 'GInteractions', matched = TRUE,
                          method = 'rejection',
                          replace = FALSE)

#> MatchedGInteractions object with 500 interactions and 3 metadata columns:
#>         seqnames1     ranges1     seqnames2     ranges2 |  feature1  feature2
#>             <Rle>   <IRanges>         <Rle>   <IRanges> | <logical> <numeric>
#>     [1]      chr1   4191-4290 ---      chr1   4191-4290 |     FALSE   5.60013
#>     [2]      chr1   3668-3767 ---      chr1   3668-3767 |     FALSE   5.43395
#>     [3]      chr1 10236-10335 ---      chr1 10236-10335 |     FALSE   4.47920
#>     [4]      chr1   8451-8550 ---      chr1   8451-8550 |     FALSE   1.45446
#>     [5]      chr1   5658-5757 ---      chr1   5658-5757 |     FALSE   2.99545
#>     ...       ...         ... ...       ...         ... .       ...       ...
#>   [496]      chr1   7472-7571 ---      chr1   7472-7571 |     FALSE   4.90340
#>   [497]      chr1   8316-8415 ---      chr1   8316-8415 |     FALSE   9.54752
#>   [498]      chr1   8987-9086 ---      chr1   8987-9086 |     FALSE   4.64822
#>   [499]      chr1   8288-8387 ---      chr1   8288-8387 |     FALSE   1.82828
#>   [500]      chr1     868-967 ---      chr1     868-967 |     FALSE   7.02681
#>            feature3
#>         <character>
#>     [1]           c
#>     [2]           c
#>     [3]           b
#>     [4]           c
#>     [5]           c
#>     ...         ...
#>   [496]           c
#>   [497]           b
#>   [498]           c
#>   [499]           d
#>   [500]           b
#>   -------
#>   regions: 10500 ranges and 0 metadata columns
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths