Generate a covariate-matched control set of ranges

matchRanges() uses a propensity score-based method to generate a covariate-matched control set of DataFrame, GRanges, or GInteractions objects.

matchRanges(focal, pool, covar, method = "rejection", replace = FALSE, ...)

# S4 method for DF_OR_df_OR_dt,DF_OR_df_OR_dt,formula,character_OR_missing,logical_OR_missing
matchRanges(focal, pool, covar, method, replace)

# S4 method for GRanges,GRanges,formula,character_OR_missing,logical_OR_missing
matchRanges(focal, pool, covar, method, replace)

# S4 method for GInteractions,GInteractions,formula,character_OR_missing,logical_OR_missing
matchRanges(focal, pool, covar, method, replace)

Arguments

focal
A DataFrame, GRanges, or GInteractions object containing the focal data to match.

pool
A DataFrame, GRanges, or GInteractions object containing the pool from which to select matches.

covar
A rhs formula with covariates on which to match.

method
A character describing which matching method to use. supported options are either 'nearest', 'rejection', or 'stratified'.

replace
TRUE/FALSE describing whether to select matches with or without replacement.

...
Additional arguments.

Value

A covariate-matched control set of data.

Details

Available inputs for focal and pool include data.frame, data.table, DataFrame, GRanges, or GInteractions. data.frame and data.table inputs are coerced to DataFrame objects and returned as MatchedDataFrame while GRanges and GInteractions objects are returned as MatchedGRanges or MatchedGInteractions, respectively.

Matching methods

method = 'nearest': Nearest neighbor matching with replacement. Finds the nearest neighbor by using a rolling join with data.table. Matching without replacement is not currently supported.

method = 'rejection': (Default) Rejection sampling with or without replacement. Uses a probability-based approach to select options in the pool that match the focal distribition.

method = 'stratified': Iterative stratified sampling with or without replacement. Bins focal and pool propensity scores by value and selects matches within bins until all focal items have a corresponding match in pool.

Examples

## Match with DataFrame
set.seed(123)
x <- makeExampleMatchedDataSet(type = 'DataFrame')
matchRanges(focal = x[x$feature1,],
            pool = x[!x$feature1,],
            covar = ~feature2 + feature3)

#> MatchedDataFrame with 500 rows and 3 columns
#>      feature1  feature2    feature3
#>     <logical> <numeric> <character>
#> 1       FALSE   5.60013           c
#> 2       FALSE   5.43395           c
#> 3       FALSE   4.47920           b
#> 4       FALSE   1.45446           c
#> 5       FALSE   2.99545           c
#> ...       ...       ...         ...
#> 496     FALSE   4.90340           c
#> 497     FALSE   9.54752           b
#> 498     FALSE   4.64822           c
#> 499     FALSE   1.82828           d
#> 500     FALSE   7.02681           b

## Match with GRanges
set.seed(123)
x <- makeExampleMatchedDataSet(type = "GRanges")
matchRanges(focal = x[x$feature1,],
            pool = x[!x$feature1,],
            covar = ~feature2 + feature3)

#> MatchedGRanges object with 500 ranges and 3 metadata columns:
#>         seqnames      ranges strand |  feature1  feature2    feature3
#>            <Rle>   <IRanges>  <Rle> | <logical> <numeric> <character>
#>     [1]     chr1   4191-4290      * |     FALSE   5.60013           c
#>     [2]     chr1   3668-3767      * |     FALSE   5.43395           c
#>     [3]     chr1 10236-10335      * |     FALSE   4.47920           b
#>     [4]     chr1   8451-8550      * |     FALSE   1.45446           c
#>     [5]     chr1   5658-5757      * |     FALSE   2.99545           c
#>     ...      ...         ...    ... .       ...       ...         ...
#>   [496]     chr1   7472-7571      * |     FALSE   4.90340           c
#>   [497]     chr1   8316-8415      * |     FALSE   9.54752           b
#>   [498]     chr1   8987-9086      * |     FALSE   4.64822           c
#>   [499]     chr1   8288-8387      * |     FALSE   1.82828           d
#>   [500]     chr1     868-967      * |     FALSE   7.02681           b
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Match with GInteractions
set.seed(123)
x <- makeExampleMatchedDataSet(type = "GInteractions")
matchRanges(focal = x[x$feature1,],
            pool = x[!x$feature1,],
            covar = ~feature2 + feature3)

#> MatchedGInteractions object with 500 interactions and 3 metadata columns:
#>         seqnames1     ranges1     seqnames2     ranges2 |  feature1  feature2
#>             <Rle>   <IRanges>         <Rle>   <IRanges> | <logical> <numeric>
#>     [1]      chr1   4191-4290 ---      chr1   4191-4290 |     FALSE   5.60013
#>     [2]      chr1   3668-3767 ---      chr1   3668-3767 |     FALSE   5.43395
#>     [3]      chr1 10236-10335 ---      chr1 10236-10335 |     FALSE   4.47920
#>     [4]      chr1   8451-8550 ---      chr1   8451-8550 |     FALSE   1.45446
#>     [5]      chr1   5658-5757 ---      chr1   5658-5757 |     FALSE   2.99545
#>     ...       ...         ... ...       ...         ... .       ...       ...
#>   [496]      chr1   7472-7571 ---      chr1   7472-7571 |     FALSE   4.90340
#>   [497]      chr1   8316-8415 ---      chr1   8316-8415 |     FALSE   9.54752
#>   [498]      chr1   8987-9086 ---      chr1   8987-9086 |     FALSE   4.64822
#>   [499]      chr1   8288-8387 ---      chr1   8288-8387 |     FALSE   1.82828
#>   [500]      chr1     868-967 ---      chr1     868-967 |     FALSE   7.02681
#>            feature3
#>         <character>
#>     [1]           c
#>     [2]           c
#>     [3]           b
#>     [4]           c
#>     [5]           c
#>     ...         ...
#>   [496]           c
#>   [497]           b
#>   [498]           c
#>   [499]           d
#>   [500]           b
#>   -------
#>   regions: 10500 ranges and 0 metadata columns
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths

## Nearest neighbor matching with replacement
set.seed(123)
x <- makeExampleMatchedDataSet(type = 'DataFrame')
matchRanges(focal = x[x$feature1,],
            pool = x[!x$feature1,],
            covar = ~feature2 + feature3,
            method = 'nearest',
            replace = TRUE)

#> MatchedDataFrame with 500 rows and 3 columns
#>      feature1  feature2    feature3
#>     <logical> <numeric> <character>
#> 1       FALSE   2.87088           c
#> 2       FALSE   3.54290           c
#> 3       FALSE   7.11436           c
#> 4       FALSE  10.78965           b
#> 5       FALSE   4.25960           c
#> ...       ...       ...         ...
#> 496     FALSE  0.173349           e
#> 497     FALSE  4.362421           a
#> 498     FALSE  3.182474           e
#> 499     FALSE  4.688994           d
#> 500     FALSE  5.068635           d

## Rejection sampling without replacement
set.seed(123)
x <- makeExampleMatchedDataSet(type = 'DataFrame')
matchRanges(focal = x[x$feature1,],
            pool = x[!x$feature1,],
            covar = ~feature2 + feature3,
            method = 'rejection',
            replace = FALSE)

#> MatchedDataFrame with 500 rows and 3 columns
#>      feature1  feature2    feature3
#>     <logical> <numeric> <character>
#> 1       FALSE   5.60013           c
#> 2       FALSE   5.43395           c
#> 3       FALSE   4.47920           b
#> 4       FALSE   1.45446           c
#> 5       FALSE   2.99545           c
#> ...       ...       ...         ...
#> 496     FALSE   4.90340           c
#> 497     FALSE   9.54752           b
#> 498     FALSE   4.64822           c
#> 499     FALSE   1.82828           d
#> 500     FALSE   7.02681           b

## Stratified sampling without replacement
set.seed(123)
x <- makeExampleMatchedDataSet(type = 'DataFrame')
matchRanges(focal = x[x$feature1,],
            pool = x[!x$feature1,],
            covar = ~feature2 + feature3,
            method = 'stratified',
            replace = FALSE)

#> MatchedDataFrame with 500 rows and 3 columns
#>      feature1  feature2    feature3
#>     <logical> <numeric> <character>
#> 1       FALSE   2.86662           c
#> 2       FALSE   3.54290           c
#> 3       FALSE   7.11436           c
#> 4       FALSE  10.78965           b
#> 5       FALSE   4.25960           c
#> ...       ...       ...         ...
#> 496     FALSE   1.32077           b
#> 497     FALSE   4.35726           a
#> 498     FALSE   4.32895           b
#> 499     FALSE   4.68899           d
#> 500     FALSE   5.06864           d